pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oxoria"]

[project]
name = "oxoria"
version = "2026.1.0"
dependencies = [
    "PySide6",
    "faiss-cpu",
    "torch",
    "transformers",
    "numpy",
    "optimum[onnxruntime]", 
    "pillow",
    "ImageHash",
    "psutil",
    "setproctitle"
]
requires-python = ">=3.10"
authors = [
  {name = "shinonome-MiDUki", email = "shiinaayame3@gmail.com"}
]
maintainers = [
  {name = "shinonome-MiDUki", email = "shiinaayame3@gmail.com"}
]
description = "Stand on the shoulders of giants"
readme = "README.md"
license = "Apache-2.0"
keywords = ["oxoria", "reference", "python"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python"
]

[project.urls]
Documentation = "https://github.com/shinonome-MiDUki/oxoria"
Repository = "https://github.com/shinonome-MiDUki/oxoria.git"
Changelog = "https://github.com/shinonome-MiDUki/oxoria/changelog.txt"

[project.scripts]
oxoria = "oxoria.__main__:main"
